src/old/pyproject.toml
[project.urls]
homepage = "https://github.com/hu-po/tatbot"
issues = "https://github.com/hu-po/tatbot/issues"
discord = "https://discord.gg/pPAFwndTJd"

[project]
name = "tatbot"
version = "0.4.1"
description="tattoo robot"
authors = [
    { name = "Hugo Ponte", email = "hu.po.xyz@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = "==3.11.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "opencv-python==4.11.0.86",
    "Pillow>=9.0.0,<11.0.0",
    "pyyaml==6.0.2",
    "safetensors==0.5.3",
    "tyro==0.9.24",
    "dacite==1.9.2",
    "paramiko==3.5.1",
    "mcp[cli]==1.9.4",
    "psutil==7.0.0",
    "pyroki @ git+https://github.com/chungmin99/pyroki.git@main#egg=pyroki",
]

[project.optional-dependencies]
bot = [
    "lerobot[tatbot,intelrealsense,smolvla] @ git+https://github.com/hu-po/lerobot.git@main",
    "pyrealsense2==2.55.1.6486",
    "trossen-arm==1.8.3",
]
gen = [
    "jax[cuda12]>=0.4.0,<0.5.0",
    "jaxtyping>=0.2.25,<1.0.0",
    "jaxlie>=1.3.4,<2.0.0",
    "networkx==3.5",
    "replicate==1.0.7",
    "scikit-image==0.25.2",
    "svgpathtools==1.7.1",
    "cairosvg==2.8.2",
]
tag = [
    "ffmpeg-python==0.2.0",
    "pupil-apriltags==1.0.4.post11",
]
viz = [
    "viser==0.2.23",
    "pyliblzfse>=0.4.1",
]
vla = [
    "lerobot[tatbot,smolvla] @ git+https://github.com/hu-po/lerobot.git@main",
]

[tool.setuptools]
py-modules = ['_bot', '_ik', '_ink', '_log', '_map', '_net', '_path', '_plan', '_scan', '_tag', '_viz']